+++ main.py
@@ -7,15 +7,19 @@ import os
 import threading
 import logging
 from flask import Flask, jsonify
+from dotenv import load_dotenv
+load_dotenv()  # 가장 먼저
 from logging_config import setup_logging
 setup_logging()
 
+"""
 # 로그 설정
 logging.basicConfig(
     level=os.environ.get("LOG_LEVEL", "INFO"),
     format="%(asctime)s [%(levelname)s] %(name)s - %(message)s",
     datefmt="%Y-%m-%d %H:%M:%S",
 )
+"""
 logger = logging.getLogger(__name__)
 
 from trader.engine import TradingEngine